Start handling transaction failures
It only handles errors that occur during the "sending" phase (i.e. user
rejects the transaction in their wallet or gas estimation fails, etc).

Still need to handle transactions that fail on the blockchain (e.g. they
run out of gas because of out-of-date hints) or disappear due to reorg.