Rename operation fields in Trove and StabilityDeposit
It is a good idea to avoid using the same name for incompatible
fields in 2 or more types that implement the same interface. This
way we avoid having to use aliases in queries that return a mix of
the 2 (or more) types.

E.g. before we'd have had to do something like:

  {
    changes(orderBy: sequenceNumber) {
      transaction {
        timestamp
      }

      ... on TroveChange {
        troveOperation: operation
      }

      ... on StabilityDepositChange {
        stabilityDepositOperation: operation
      }
    }
  }

Whereas now it's simply:

  {
    changes(orderBy: sequenceNumber) {
      transaction {
        timestamp
      }

      ... on TroveChange {
        troveOperation
      }

      ... on StabilityDepositChange {
        stabilityDepositOperation
      }
    }
  }