Add new package for launchkit-wizard-app
This will be one of the components of the LaunchKit wizard. It'll be
a web app that runs locally and can be used to configure the
LaunchKit frontend.

Importantly, it will take care of registering the frontend operator's
Ethereum address as a Liquity frontend. This is why we opted to
implement the configurator as a web app, so we can rely on MetaMask
as a wallet.