contracts: Address follow-up audit issues related to Unipool
- TOB-LQT-016: Renamed `exit` to `withdrawAndClaim` to avoid
confusion.
We want to keep `claimReward` reverting if there’s no reward
available, as we think it makes sense and it’s more consistent with
the rest of the system, where liquidations and redemptions revert in
case of zero final result.
The renamed function is just a wrapper to call 2 functions in one
transaction, so if one of them reverts, the other one must revert
too. The frontend can easily handle this and redirect users to call
only withdraw if there are no rewards available. It wouldn’t make much
sense to call the wrapper if only one function is needed, it would
just add complexity and waste gas.

- Address the comments in “C. Code Quality” related to Unipool
contract too.