feat: truncate redemption amount to ensure maximum possible redemption
With the new minimum debt requirement for Troves, a redemption that would
leave the last Trove of the sequence with less debt than the minimum is no
longer possible. Such a redemption attempt has to leave the last Trove
completely untouched, which can result in a significantly lower redeemed
amount than originally planned.

To avoid this, we truncate such redemptions so that they leave the last
Trove with exactly the minimum debt.