fix: eliminate bogus `TroveChange` upon capped liquidation
The issue was introduced when capped liquidations were added to the backend.
Up until then, the `_coll` and `_debt` parameters of `TroveLiquidated` events
reflected the final state of the Trove immediately before the liquidation.

The subgraph relied on this to insert an `accrueRewards` operation in case
there have been redistributions between the last Trove adjustment and the
liquidation. However, in the case of a capped liquidation, `_coll` now reflects
the capped portion of the collateral that's liquidated, so we can no longer
rely on the event's parameters.

Instead, we should calculate the redistribution from the Trove's snapshots,
stake and the global redistribution counters.

Fixes #580.